# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: StreamHub - Community TV Streaming Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /channels:
    get:
      operationId: listChannels
      tags: [channels]
      summary: List all channels
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Channel list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Channel"
    post:
      operationId: createChannel
      tags: [channels]
      summary: Create a new channel
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChannelRequest"
      responses:
        "201":
          description: Channel created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
  /channels/{id}:
    get:
      operationId: getChannel
      tags: [channels]
      summary: Get channel by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Channel details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateChannel
      tags: [channels]
      summary: Update a channel
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateChannelRequest"
      responses:
        "200":
          description: Channel updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
    delete:
      operationId: deleteChannel
      tags: [channels]
      summary: Delete a channel
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Channel deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /channels/{id}/regenerate-key:
    post:
      operationId: regenerateStreamKey
      tags: [channels]
      summary: Regenerate stream key for a channel
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: New stream key generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
  /streams/active:
    get:
      operationId: getActiveStreams
      tags: [streams]
      summary: Get active streams from SRS
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Active stream list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActiveStream"
  /streams/clients:
    get:
      operationId: getStreamClients
      tags: [streams]
      summary: Get active clients from SRS
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Client list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreamClients"
  /streams/bandwidth:
    get:
      operationId: getStreamBandwidth
      tags: [streams]
      summary: Get bandwidth stats from SRS
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Bandwidth statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BandwidthStats"
  /streams/stats:
    get:
      operationId: getStreamStats
      tags: [streams]
      summary: Get aggregated stream stats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Aggregated statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Channel:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        streamKey:
          type: string
        rtmpUrl:
          type: string
        hlsUrl:
          type: string
        webrtcUrl:
          type: string
        isActive:
          type: boolean
        createdById:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - streamKey
        - rtmpUrl
        - hlsUrl
        - webrtcUrl
        - isActive
        - createdById
        - createdAt
        - updatedAt
    CreateChannelRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
      required:
        - name
    UpdateChannelRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        isActive:
          type: boolean
    ActiveStream:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        channelName:
          type: string
        clientId:
          type: string
        vhost:
          type: string
        app:
          type: string
        stream:
          type: string
        bitrate:
          type: number
        viewers:
          type: integer
        uptime:
          type: number
        isOnline:
          type: boolean
        videoCodec:
          type: string
        audioCodec:
          type: string
        width:
          type: integer
        height:
          type: integer
      required:
        - id
        - name
        - isOnline
    StreamClients:
      type: object
      properties:
        totalClients:
          type: integer
        players:
          type: integer
        publishers:
          type: integer
        clients:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              ip:
                type: string
              type:
                type: string
              url:
                type: string
              alive:
                type: number
      required:
        - totalClients
        - players
        - publishers
    BandwidthStats:
      type: object
      properties:
        ingressBps:
          type: number
        egressBps:
          type: number
        ingressBytes:
          type: number
        egressBytes:
          type: number
      required:
        - ingressBps
        - egressBps
        - ingressBytes
        - egressBytes
    DashboardStats:
      type: object
      properties:
        totalChannels:
          type: integer
        activeStreams:
          type: integer
        totalViewers:
          type: integer
        totalBandwidth:
          type: number
        totalUsers:
          type: integer
        totalRevenue:
          type: number
        recentViewerCounts:
          type: array
          items:
            type: object
            properties:
              timestamp:
                type: string
              count:
                type: integer
      required:
        - totalChannels
        - activeStreams
        - totalViewers
        - totalBandwidth
        - totalUsers
        - totalRevenue
...